backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
google-generativeai
supabase
anthropic
pillow
aiohttp
